Move ld scripts into esp-hal-common
