+++ esp-hal-common/build.rs
@@ -148,7 +148,22 @@ fn main() {
     if cfg!(feature = "esp32") || cfg!(feature = "esp32s2") || cfg!(feature = "esp32s3") {
         fs::copy("ld/xtensa/hal-defaults.x", out.join("hal-defaults.x")).unwrap();
         fs::copy("ld/xtensa/rom.x", out.join("alias.x")).unwrap();
+        copy_dir_all("ld/sections", out).unwrap(); // TODO once supported on riscv move outside of if
     } else {
         fs::copy("ld/riscv/hal-defaults.x", out.join("hal-defaults.x")).unwrap();
     }
+}
+
+fn copy_dir_all(src: impl AsRef<std::path::Path>, dst: impl AsRef<std::path::Path>) -> std::io::Result<()> {
+    fs::create_dir_all(&dst)?;
+    for entry in fs::read_dir(src)? {
+        let entry = entry?;
+        let ty = entry.file_type()?;
+        if ty.is_dir() {
+            copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else {
+            fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        }
+    }
+    Ok(())
 }
