+++ esp-hal-common/ld/sections/external.x
