+++ esp-hal-common/ld/sections/rodata.x
