+++ esp-hal-common/ld/sections/rtc_fast.x
