+++ esp-hal-common/ld/sections/rtc_slow.x
