+++ esp-hal-common/ld/sections/rwdata.x
