+++ esp-hal-common/ld/sections/rwtext.x
