+++ esp-hal-common/ld/sections/text.x
